Feat: migration (#26)
* feat: simple migration policy
* refactor: use IsPodReady function
* feat: prevent dispatcher to route requests to not ready pods
* feat: tons of shitty things. read description
* feat: stop sync only for not ready functions
* feat: fix termination period
* fix: increase maximum delay upper bound
* feat: add proxy metric table
* fix: delete unused const query string
* refactor: extract scheduler name in const field
* refactor: extract default termination grace period
* feat: change latency to float and fix backend update
* refactor: user-fn.yaml
* refactor: manifests used in experiments

+++ pkg/system-controller/pkg/delayclient/client.go
@@ -11,7 +11,7 @@ import (
 type NodeDelay struct {
 	FromNode string
 	ToNode   string
-	Latency  int
+	Latency  float64
 }
 
 const (
